Starting work on image filtering
So far I've migrated the main color buffer used by the render target to
an atomically updated buffer of floating point colors and do still get
the same result out when pretending to use a box filter basically.

Next work will be actually implementing a box filter and making use of
it as described in PBR and then implementing some other filters they
discuss and seeing if everything explodes or not. Fun!