Working on hunting a bug causing a crash in model madness w/ dof
I think it may be out of bounds writes to the float buf since it's an
access violation and we don't do bounds checking there but unsure.
Comitting a potential fix but haven't tested it thoroughly yet. Would
really like to run the whole program through valgrind but it takes so
long :(